move(MenuGroup): move MenuGroup to Menu
