+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -4,8 +4,8 @@ import kitchenpos.application.FakePurgomalumClient;
 import kitchenpos.product.infra.InMemoryProductRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
-import kitchenpos.menugroup.infra.InMemoryMenuGroupRepository;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menu.infra.InMemoryMenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
